refactor: update services, APIs and DBs exporting structure
