+++ src/lib/services/API/index.ts
@@ -1,0 +1,2 @@
+export * from "./NewMessagesAPI";
+export * from "./UsersAPI";
