+++ src/lib/services/index.ts
@@ -1,4 +1,4 @@
 export * from "./AuthenticationAPI";
-export * from "./MessagesAPI";
+export * from "./API";
 export * from "./RoomsAPI";
 export * from "./QuestionsAPI";
